requirements.txt
# Apify SDK
apify>=2.0.0

# HTTP client
httpx>=0.27.0

# AI Platform SDKs
openai>=1.0.0
anthropic>=0.40.0
google-generativeai>=0.8.0

# Data validation
pydantic>=2.0.0
